+++ integration/test/Testlib/ModService.hs
@@ -242,7 +242,8 @@ startDynamicBackend resource beOverrides = do
           gundeckCfg = setField "cassandra.keyspace" resource.berGundeckKeyspace,
           backgroundWorkerCfg =
             setField "cassandra.keyspace" resource.berGundeckKeyspace
-              >=> setField "cassandraGalley.keyspace" resource.berGalleyKeyspace,
+              >=> setField "cassandraGalley.keyspace" resource.berGalleyKeyspace
+              >=> setField "cassandraBrig.keyspace" resource.berBrigKeyspace,
           cannonCfg =
             setField "cassandra.keyspace" resource.berGundeckKeyspace
         }
@@ -645,12 +646,12 @@ prepareNginzRuntimeFiles resource = do
   sm <- getServiceMap domain
   mBaseDir <- asks (.servicesCwdBase)
   case mBaseDir of
-    Nothing -> liftIO $ prepareNginzK8sRuntimeFiles domain sm
+    Nothing -> liftIO $ prepareNginzK8sRuntimeFiles sm
     Just basedir -> liftIO $ prepareNginzLocalRuntimeFiles resource sm basedir
 
-prepareNginzK8sRuntimeFiles :: String -> ServiceMap -> IO (FilePath, FilePath, FilePath)
-prepareNginzK8sRuntimeFiles domain sm = do
-  tmpDir <- createTempDirectory "/tmp" ("nginz" <> "-" <> domain)
+prepareNginzK8sRuntimeFiles :: ServiceMap -> IO (FilePath, FilePath, FilePath)
+prepareNginzK8sRuntimeFiles sm = do
+  tmpDir <- createTempDirectory "/tmp" "nginz-"
   copyDirectoryRecursively "/etc/wire/nginz/" tmpDir
 
   let nginxConfFile = tmpDir </> "conf" </> "nginx.conf"
@@ -669,12 +670,10 @@ prepareNginzK8sRuntimeFiles domain sm = do
 
 prepareNginzLocalRuntimeFiles :: BackendResource -> ServiceMap -> FilePath -> IO (FilePath, FilePath, FilePath)
 prepareNginzLocalRuntimeFiles resource sm basedir = do
-  let domain = berDomain resource
-
   -- Create a whole temporary directory and copy all nginx's config files.
   -- This is necessary because nginx assumes local imports are relative to
   -- the location of the main configuration file.
-  tmpDir <- createTempDirectory "/tmp" ("nginz" <> "-" <> domain)
+  tmpDir <- createTempDirectory "/tmp" "nginz-"
 
   -- copy all config files into the tmp dir
   let from = basedir </> "nginz" </> "integration-test"